Extracts dumber status particular presenters for auctions
        This refactor converts repeated conditionals around auction
        status into fairly declarative polymorphic presenters. The
        twitter data types and labels will also move in these classes
        in the next commits.
