+++ app/models/presenter/auction_status/future.rb
@@ -1,0 +1,17 @@
+module Presenter
+  module AuctionStatus
+    class Future < Struct.new(:auction)
+      def label_class
+        'auction-label-future'
+      end
+
+      def label
+        'Coming Soon'
+      end
+
+      def status_tag_data
+        auction.human_start_time
+      end
+    end
+  end
+end
